LU-3319 procfs: provide framework for seq_file handling
In linux 3.10 kernels and above the proc file system interface
changed. To handle this change we need to move from the current
method of using proc_read and proc_write function pointers to
using the seq_file. This patch provides the base to enable moving
all the lustre proc handling gradually over to this new method
of proc data handling.

Reviewed-on: http://review.whamcloud.com/7135
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>

+++ lustre/ptlrpc/nrs_crr.c
@@ -718,7 +718,7 @@ static int ptlrpc_lprocfs_wr_nrs_crrn_quantum(struct file *file,
 	long			     quantum_reg;
 	long			     quantum_hp;
 	/** lprocfs_find_named_value() modifies its argument, so keep a copy */
-	unsigned long		     count_copy;
+	size_t			     count_copy;
 	int			     rc = 0;
 	int			     rc2 = 0;
